FIX 동시성 문제 해결 로직 수정
