+++ src/test/java/com/jeontongju/coupon/CouponApplicationTests.java
@@ -8,6 +8,8 @@ import com.jeontongju.coupon.service.CouponService;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+
+import com.jeontongju.coupon.utils.ReceiveManager;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,6 +21,7 @@ class CouponApplicationTests {
   @Autowired private CouponService couponService;
 
   @Autowired private RedissonLockCouponFacade redissonLockCouponFacade;
+  @Autowired private ReceiveManager receiveManager;
 
   @Test
   @DisplayName("프로모션 쿠폰 수령 시, 동시에 요청이 들어오면 Race Condition 문제가 생긴다.")
@@ -34,7 +37,7 @@ class CouponApplicationTests {
       executorService.submit(
           () -> {
             try {
-              redissonLockCouponFacade.decreasePromotionCoupon("v5F5-4125-WXHz", 1L);
+              receiveManager.decreasePromotionCoupon("v5F5-4125-WXHz", 1L);
             } finally {
               latch.countDown();
             }
